DRTII-1213 Render performance fixes
