+++ client/src/main/scala/drt/client/components/NationalityFlaggingComponent.scala
@@ -9,26 +9,24 @@ import io.kinoplan.scalajs.react.material.ui.core.system.SxProps
 import io.kinoplan.scalajs.react.material.ui.icons.MuiIcons
 import io.kinoplan.scalajs.react.material.ui.icons.MuiIconsModule.ExpandMore
 import japgolly.scalajs.react.component.Scala.Component
-import japgolly.scalajs.react.extra.ReusabilityOverlay
 import japgolly.scalajs.react.facade.React.Node
 import japgolly.scalajs.react.vdom.html_<^._
 import japgolly.scalajs.react.{CtorType, _}
 
 import scala.scalajs.js
-import scala.scalajs.js.JSConverters.JSRichIterableOnce
 
 
 object NationalityFlaggingComponent {
   case class Props(flaggedNationalities: Set[Country])
 
-  case class State(flaggedNationalities: Set[Country], inputValue: String)
+  case class State(inputValue: String)
 
-  implicit val propsReuse: Reusability[Props] = Reusability.always
+  implicit val propsReuse: Reusability[Props] = Reusability.derive[Props]
   implicit val stateReuse: Reusability[State] = Reusability.derive[State]
 
   val component: Component[Props, State, Unit, CtorType.Props] = ScalaComponent.builder[Props]("nationality-flagger")
-    .initialStateFromProps(p => State(p.flaggedNationalities, ""))
-    .renderS { (scope, state) =>
+    .initialState(State(""))
+    .renderPS { (scope, props, state) =>
       val acTextInput: js.Function1[AutocompleteRenderInputParams, Node] = (params: AutocompleteRenderInputParams) => {
         MuiTextField(
           id = params.id,
@@ -45,7 +43,7 @@ object NationalityFlaggingComponent {
         MuiAutocompleteOption(label = s"${c.name} (${c.threeLetterCode})", id = c.id)
       }: _*)
 
-      val flagCount = state.flaggedNationalities.size
+      val flagCount = props.flaggedNationalities.size
       val flagCountDisplay = flagCount match {
         case 0 => ""
         case n => s" ($n)"
@@ -56,8 +54,14 @@ object NationalityFlaggingComponent {
           <.div(
             ^.style := js.Dictionary("display" -> "inline"),
             MuiButton(variant = MuiButton.Variant.text)(
-              "Clear all",
-              ^.onClick --> Callback(SPACircuit.dispatch(ClearFlaggedNationalities))
+              MuiTypography(variant = "body1", sx = SxProps(Map("textDecoration" -> "underline", "fontWeight" -> "bold")))(
+                "Clear all"
+              ),
+              ^.onClick ==> { e => {
+                e.preventDefault()
+                Callback(SPACircuit.dispatch(ClearFlaggedNationalities))
+              }
+              }
             ),
           )
       }
@@ -65,8 +69,10 @@ object NationalityFlaggingComponent {
         ^.style := js.Dictionary("display" -> "flex"),
         MuiAccordion(elevation = 0)(
           MuiAccordionSummary(expandIcon = toVdomNode(MuiIcons(ExpandMore)()))(
-            MuiTypography(variant = "body1", sx = SxProps(Map("textDecoration" -> "underline", "fontWeight" -> "bold")))(
-              s"Flag flights $flagCountDisplay",
+            <.div(^.style := js.Dictionary("display" -> "flex", "alignItems" -> "center", "gap" -> "16px"),
+              MuiTypography(variant = "body1", sx = SxProps(Map("textDecoration" -> "underline", "fontWeight" -> "bold")))(
+                s"Flag flights $flagCountDisplay"
+              ),
               clearFlags
             )
           ),
@@ -83,52 +89,25 @@ object NationalityFlaggingComponent {
                 },
                 inputValue = state.inputValue,
                 onInputChange = (_: ReactEvent, value: String) => scope.modState(_.copy(inputValue = value)),
-//                value = null,
+                value = null,
                 onChange = (_: ReactEvent, value: MuiAutocompleteOption) => value match {
                   case option: MuiAutocompleteOption =>
                     CountryOptions.countries.find(_.id == option.id) match {
                       case Some(c) =>
-                        scope.modState(_.copy(flaggedNationalities = state.flaggedNationalities + c)).map(_ =>
-                          SPACircuit.dispatch(AddFlaggedNationality(c))
-                        )
+                        scope.modState(_.copy(inputValue = ""))
+                          .map(_ => SPACircuit.dispatch(AddFlaggedNationality(c)))
                       case None => Callback.empty
                     }
 
                   case _ => Callback.empty
                 },
               )(),
-//              MuiAutocomplete[String](
-//                disablePortal = true,
-//                options = CountryOptions.countries.map(c => s"${c.name} (${c.threeLetterCode})").toJSArray,
-//                renderInput = acTextInput,
-//                sx = SxProps(Map("minWidth" -> "250px")),
-//                getOptionLabel = (o: MuiAutocompleteOption) => o.label,
-//                isOptionEqualToValue = (o1: MuiAutocompleteOption, o2: MuiAutocompleteOption) => {
-//                  o1.id == o2.id
-//                },
-//                value = null,
-//                onChange = (_: ReactEvent, value: MuiAutocompleteOption) => value match {
-//                  case option: MuiAutocompleteOption =>
-//                    CountryOptions.countries.find(_.id == option.id) match {
-//                      case Some(c) =>
-//                        scope.modState(_.copy(flaggedNationalities = state.flaggedNationalities + c)).map(_ =>
-//                          SPACircuit.dispatch(AddFlaggedNationality(c))
-//                        )
-//                      case None => Callback.empty
-//                    }
-//
-//                  case _ => Callback.empty
-//                },
-//              )(),
               <.div(
                 ^.style := js.Dictionary("display" -> "flex", "alignItems" -> "center", "gap" -> "16px"),
-                state.flaggedNationalities.toList
+                props.flaggedNationalities.toList
                   .sortBy(_.name)
                   .map { c =>
-                    val onDelete = (_: ReactEvent) =>
-                      scope.modState(_.copy(flaggedNationalities = state.flaggedNationalities - c)).map(_ =>
-                        SPACircuit.dispatch(RemoveFlaggedNationality(c))
-                      )
+                    val onDelete = (_: ReactEvent) => Callback(SPACircuit.dispatch(RemoveFlaggedNationality(c)))
                     MuiChip(label = s"${c.name} (${c.threeLetterCode})".toVdom, onDelete = onDelete)()
                   }
                   .toTagMod
@@ -139,7 +118,7 @@ object NationalityFlaggingComponent {
       )
     }
     .configure(Reusability.shouldComponentUpdate)
-//    .configure(ReusabilityOverlay.install)
+    //    .configure(ReusabilityOverlay.install)
     .build
 }
