+++ server/src/main/assets/stylesheets/main.less
@@ -1196,9 +1196,9 @@ p {
 
 .flex-horz-between {
   display: flex;
-  width: 100%;
   flex-wrap: wrap;
   justify-content: space-between;
+  flex-grow: 1;
 }
 
 .terminal-content-header {
@@ -2001,7 +2001,7 @@ th h2 {
 }
 
 #arrivals table thead {
-  background-color: #2989C6;
+  background-color: #335096;
   color: #fff;
 }
